requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0